Synch Now Works With Seeking
Synching is now starting to be complete. There are a few problems with
jitters when seeking and playing and the timing is about half a second
off, but for the most part it should work. There appears to be a bug
where seeking pauses some clients and seems to be much more common when
adjusting client side instead of host.